+++ src/fields/data/SingleOptionFieldData.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace amici\SuperDynamicFields\fields\data;
+
+class SingleOptionFieldData extends OptionData
+{
+
+    private $_options = [];
+
+    public function getOptions(): array
+    {
+        return $this->_options;
+    }
+
+    public function setOptions(array $options)
+    {
+        $this->_options = $options;
+    }
+
+}